fix: adding suggested fixes, testing fixes soon
