+++ tests/conftest.py
@@ -3,9 +3,7 @@ import os
 import os.path
 from pathlib import Path
 import re
-import subprocess
 import sys
-import time
 
 import pytest
 
@@ -52,47 +50,6 @@ def getAssetPath():
         return path
 
     return loader
-
-
-def checkCarlaPath():
-    CARLA_ROOT = os.environ.get("CARLA_ROOT")
-    if not CARLA_ROOT:
-        pytest.skip("CARLA_ROOT env variable not set.")
-    return CARLA_ROOT
-
-
-@pytest.fixture
-def launchCarlaServer():
-    CARLA_ROOT = checkCarlaPath()
-    carla_process = subprocess.Popen(
-        f"bash {CARLA_ROOT}/CarlaUE4.sh -RenderOffScreen", shell=True
-    )
-    # NOTE: CARLA server takes time to start up
-    time.sleep(3)
-    yield
-    while carla_process.poll() is None:
-        carla_process.kill()
-        time.sleep(0.1)
-    # NOTE: CARLA server hangs. The first `kill` usually works,
-    # but for safe measure we needed to kill hanging zombie processes
-    for _ in range(10):
-        carla_process.kill()
-        time.sleep(0.1)
-    time.sleep(5)
-
-
-@pytest.fixture
-def getCarlaSimulator():
-    from scenic.simulators.carla import CarlaSimulator
-
-    base = Path(__file__).parent.parent / "assets" / "maps" / "CARLA"
-
-    def _getCarlaSimulator(town):
-        path = os.path.join(base, town + ".xodr")
-        simulator = CarlaSimulator(carla_map=town, map_path=path)
-        return (simulator, town, path)
-
-    return _getCarlaSimulator
 
 
 ## Command-line options
